Move user activity migrations to 1.9.9
