+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/mysql/v199/Migration.java
@@ -1,10 +1,10 @@
-package org.openmetadata.service.migration.postgres.v198;
+package org.openmetadata.service.migration.mysql.v199;
 
 import lombok.SneakyThrows;
 import org.openmetadata.service.jdbi3.locator.ConnectionType;
 import org.openmetadata.service.migration.api.MigrationProcessImpl;
 import org.openmetadata.service.migration.utils.MigrationFile;
-import org.openmetadata.service.migration.utils.v198.MigrationUtil;
+import org.openmetadata.service.migration.utils.v199.MigrationUtil;
 
 public class Migration extends MigrationProcessImpl {
 
@@ -15,7 +15,7 @@ public class Migration extends MigrationProcessImpl {
   @Override
   @SneakyThrows
   public void runDataMigration() {
-    MigrationUtil migrationUtil = new MigrationUtil(handle, ConnectionType.POSTGRES);
+    MigrationUtil migrationUtil = new MigrationUtil(handle, ConnectionType.MYSQL);
     migrationUtil.addUserActivityColumns();
   }
 }
